+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### 20200429
+- Added option to import neurons from workspace to `views\RenderApp.m`
+- Added `lib\uigetvar.m`
+
 ### 20200427
 - Added support for pressing `Enter` key to import neurons in `views\RenderApp.m`